New deploy target in fabfile.
We now check the server name in the deploy() target rather than having two
separate deploy / migrate steps.  This should remove an unnecessary restart of
verdant-rca on the migration server, and also ensure that we restart verdant on
rca3 after running compress on rca2, which is required to ensure the correct
manifest.json is loaded by django-compressor. (This was the cause of problems
during a deployment yesterday.)