Remove "click anywhere on a listing item to edit it" behaviour. With page permissions in place, we can't guarantee that the first link in the list item is 'edit', which makes the behaviour inconsistent.
It also led to some clunky behaviour - e.g. clicking one of the 'preview' links would open the preview in a new window, AND take you to the edit page.

To mitigate losing this behaviour, make the h2 link a block element so that you can at least click anywhere right of the title text.