Add hashtags to have links to filtered content
This modifies only two files: filters.js to update the hashtags in url
whenever the selected filters are changed in a page and site.js to read
the hashtags when the page is loaded and update the filters with the
hashtag values. Empty filter values won't be visible in the hashtags. This
needs testing.