Fix minor bug with staff feed source
If for some reason a staff does *not* have roles for the school he belongs
to theb an exception will be thrown when rendering the staff_feed_source.

This shouln't happen under normal circumstances but the behavior has been
changed to just output an empty role instead of throwing the exception.