selecting templates based on the year of the show with fallback templates in case the templates don't exist yet
With this, you can simply add templates for the specific year that you
want to change the layout for by adding "_year" (e.g. "_2016") at the
end of the template name. If such a template exists for a given year,
that template will be used, otherwise the non-numbered template will be
used.