showing new logo
As in the year before, the logo for that page is an SVG file with only
paths. If you edit the logo, please make sure that all objects are
converted to paths so the rendering does not depend on the users'
installed fonts.

Note that the six school logos are already present but not used.