showing school-specific logo bar
for each school, a specific logo bar (that only shows one of the '16's
in the logo) is available. These different images need to be selected
depending on the value of the `school` template variable. This variable
is set for school pages, but also for programme and student pages, so
the correct logo is always shown.