using year-specific templates in children pages of ShowIndex, too
This is done by a mixin `PageWithYearTemplate` that overrides the
`get_template` function of the Page model with one that looks up the
`year` value of its index page (if present). It then selects either the
template with '_year' inserted into it or, if that doesn't exist, the
regular page template.