[#830] Added hide_animation_videos flag to ShowIndexPage (#75)
https://projects.torchbox.com/projects/rca-django-cms-project/tickets/830

We need to hide animation videos in every carousel except on the private animation show pages.

We previously relied on a bit of template logic to control this, but that logic broke recently. This commit replaces this logic with a "hide_animation_videos" checkbox (default to checked) that an administrator can uncheck on pages where animation videos should be allowed.