[835] Added migration to reassign pages with a renamed programme
Addresses:
https://projects.torchbox.com/projects/rca-django-cms-project/tickets/835

RCA rename their programmes sometimes but sometimes this has been
represented as a new programme instead of renaming an existing one. The
result is we have two programme records representing the same programme.

This migration remaps any pages that use a programme with an old name to
instead use the new renamed one.

Mapping is defined in:
https://projects.torchbox.com/projects/rca-django-cms-project/tickets/835#update-37475581

To save development time, we only do this for content types that have a lot of pages. Other page types will be changed manually.

List of content types defined in:
https://projects.torchbox.com/projects/rca-django-cms-project/tickets/835#update-38772748